requirements.txt
google-generativeai==0.3.2
streamlit==1.32.2
python-dotenv
